Add Current Weather Fragment; Use response header cache

+++ app/src/main/java/eldeveloper13/weatherapp/services/darksky/ForecastResponse.java
@@ -20,7 +20,7 @@ public class ForecastResponse {
         long time;
         String summary;
         String icon;
-        int dewPoint;
+        double dewPoint;
         double humidity;
         double cloudCover;
         int nearestStormDistance;
